feat:権限によるTODO表示制限機能、TODO詳細表示機能を実装。

+++ src/main/java/com/codestep/TodoApp/entities/TodoItem.java
@@ -34,7 +34,7 @@ public class TodoItem {
 	private String title;
 	
 	@Column(nullable = false)
-	private long done;
+	private int done;
 	
 	@Column(nullable = false)
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
